Major rework of the API configs
We have an interesting challenge of having to talk to what is ultimately
many different app with the use of app_tokens. These apps can also have
multiple events. It's entirely possible that we would want the same
event tracked in all apps but they will all have different event_tokens.
It would be nice to refer to these events by string or symbols and this
is exactly what we did. If we ever have to use app_token and event_token
explicitely we can also do that.